player speed down to 5 (10 was faster then can be controled in my oppinion) Also added the func "start_over" that disabled a situation of an infinit falling
